# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Delta Higher Institute Correspondence System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResult {
  success: boolean;
  id: number;
}

export interface Department {
  id: number;
  name: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateDepartmentBody {
  name: string;
  description?: string | null;
}

export type EmployeeRole = (typeof EmployeeRole)[keyof typeof EmployeeRole];

export const EmployeeRole = {
  admin: "admin",
  supervisor: "supervisor",
  employee: "employee",
} as const;

export interface Employee {
  id: number;
  name: string;
  email: string;
  role: EmployeeRole;
  departmentId?: number | null;
  departmentName?: string | null;
  createdAt: string;
}

export type CreateEmployeeBodyRole =
  (typeof CreateEmployeeBodyRole)[keyof typeof CreateEmployeeBodyRole];

export const CreateEmployeeBodyRole = {
  admin: "admin",
  supervisor: "supervisor",
  employee: "employee",
} as const;

export interface CreateEmployeeBody {
  name: string;
  email: string;
  role: CreateEmployeeBodyRole;
  departmentId?: number | null;
}

export type CorrespondenceType =
  (typeof CorrespondenceType)[keyof typeof CorrespondenceType];

export const CorrespondenceType = {
  incoming: "incoming",
  outgoing: "outgoing",
  internal: "internal",
} as const;

export type CorrespondenceStatus =
  (typeof CorrespondenceStatus)[keyof typeof CorrespondenceStatus];

export const CorrespondenceStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  rejected: "rejected",
  archived: "archived",
} as const;

export type CorrespondencePriority =
  (typeof CorrespondencePriority)[keyof typeof CorrespondencePriority];

export const CorrespondencePriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface Correspondence {
  id: number;
  referenceNumber: string;
  subject: string;
  body?: string | null;
  type: CorrespondenceType;
  status: CorrespondenceStatus;
  priority: CorrespondencePriority;
  fromDepartmentId?: number | null;
  fromDepartmentName?: string | null;
  toDepartmentId?: number | null;
  toDepartmentName?: string | null;
  assignedToId?: number | null;
  assignedToName?: string | null;
  createdById?: number | null;
  createdByName?: string | null;
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CorrespondenceHistory {
  id: number;
  correspondenceId: number;
  action: string;
  notes?: string | null;
  performedById?: number | null;
  performedByName?: string | null;
  createdAt: string;
}

export type CorrespondenceDetail = Correspondence & {
  history?: CorrespondenceHistory[];
};

export type CreateCorrespondenceBodyType =
  (typeof CreateCorrespondenceBodyType)[keyof typeof CreateCorrespondenceBodyType];

export const CreateCorrespondenceBodyType = {
  incoming: "incoming",
  outgoing: "outgoing",
  internal: "internal",
} as const;

export type CreateCorrespondenceBodyPriority =
  (typeof CreateCorrespondenceBodyPriority)[keyof typeof CreateCorrespondenceBodyPriority];

export const CreateCorrespondenceBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateCorrespondenceBody {
  subject: string;
  body?: string | null;
  type: CreateCorrespondenceBodyType;
  priority: CreateCorrespondenceBodyPriority;
  fromDepartmentId?: number | null;
  toDepartmentId?: number | null;
  assignedToId?: number | null;
  createdById?: number | null;
  dueDate?: string | null;
}

export type UpdateCorrespondenceBodyStatus =
  (typeof UpdateCorrespondenceBodyStatus)[keyof typeof UpdateCorrespondenceBodyStatus];

export const UpdateCorrespondenceBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  rejected: "rejected",
  archived: "archived",
} as const;

export type UpdateCorrespondenceBodyPriority =
  (typeof UpdateCorrespondenceBodyPriority)[keyof typeof UpdateCorrespondenceBodyPriority];

export const UpdateCorrespondenceBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateCorrespondenceBody {
  subject?: string;
  body?: string | null;
  status?: UpdateCorrespondenceBodyStatus;
  priority?: UpdateCorrespondenceBodyPriority;
  assignedToId?: number | null;
  dueDate?: string | null;
}

export interface CreateHistoryBody {
  action: string;
  notes?: string | null;
  performedById?: number | null;
}

export interface Archive {
  id: number;
  correspondenceId: number;
  archiveNumber: string;
  archiveLocation?: string | null;
  notes?: string | null;
  archivedById?: number | null;
  archivedByName?: string | null;
  archivedAt: string;
  correspondence: Correspondence;
}

export interface ArchiveCorrespondenceBody {
  archiveLocation?: string | null;
  notes?: string | null;
  archivedById?: number | null;
}

export interface DashboardSummary {
  totalCorrespondences: number;
  pendingCount: number;
  inProgressCount: number;
  completedCount: number;
  archivedCount: number;
  urgentCount: number;
  totalDepartments: number;
  totalEmployees: number;
}

export interface DepartmentCount {
  departmentId: number;
  departmentName: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type ListEmployeesParams = {
  departmentId?: number;
};

export type ListCorrespondencesParams = {
  status?: ListCorrespondencesStatus;
  type?: ListCorrespondencesType;
  departmentId?: number;
  assignedTo?: number;
};

export type ListCorrespondencesStatus =
  (typeof ListCorrespondencesStatus)[keyof typeof ListCorrespondencesStatus];

export const ListCorrespondencesStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  rejected: "rejected",
  archived: "archived",
} as const;

export type ListCorrespondencesType =
  (typeof ListCorrespondencesType)[keyof typeof ListCorrespondencesType];

export const ListCorrespondencesType = {
  incoming: "incoming",
  outgoing: "outgoing",
  internal: "internal",
} as const;

export type ListArchiveParams = {
  departmentId?: number;
  year?: number;
};

export type GetRecentCorrespondencesParams = {
  limit?: number;
};