# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Delta Higher Institute Correspondence System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /correspondences:
    get:
      operationId: listCorrespondences
      tags: [correspondences]
      summary: List correspondences
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed, rejected, archived]
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [incoming, outgoing, internal]
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
        - name: assignedTo
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of correspondences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Correspondence"
    post:
      operationId: createCorrespondence
      tags: [correspondences]
      summary: Create a new correspondence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCorrespondenceBody"
      responses:
        "201":
          description: Created correspondence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Correspondence"
  /correspondences/{id}:
    get:
      operationId: getCorrespondence
      tags: [correspondences]
      summary: Get correspondence by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Correspondence with history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrespondenceDetail"
    put:
      operationId: updateCorrespondence
      tags: [correspondences]
      summary: Update correspondence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCorrespondenceBody"
      responses:
        "200":
          description: Updated correspondence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Correspondence"
    delete:
      operationId: deleteCorrespondence
      tags: [correspondences]
      summary: Delete correspondence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /correspondences/{id}/archive:
    post:
      operationId: archiveCorrespondence
      tags: [correspondences]
      summary: Archive a correspondence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveCorrespondenceBody"
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Archive"
  /correspondences/{id}/history:
    post:
      operationId: addCorrespondenceHistory
      tags: [correspondences]
      summary: Add a history entry to a correspondence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHistoryBody"
      responses:
        "201":
          description: History entry added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrespondenceHistory"
...
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: integer
      required:
        - success
        - id
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required:
        - name
...
    Correspondence:
      type: object
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        subject:
          type: string
        body:
          type: string
          nullable: true
        type:
          type: string
          enum: [incoming, outgoing, internal]
        status:
          type: string
          enum: [pending, in_progress, completed, rejected, archived]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        fromDepartmentId:
          type: integer
          nullable: true
        fromDepartmentName:
          type: string
          nullable: true
        toDepartmentId:
          type: integer
          nullable: true
        toDepartmentName:
          type: string
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        createdById:
          type: integer
          nullable: true
        createdByName:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - referenceNumber
        - subject
        - type
        - status
        - priority
        - createdAt
        - updatedAt
    CorrespondenceDetail:
      allOf:
        - $ref: "#/components/schemas/Correspondence"
        - type: object
          properties:
            history:
              type: array
              items:
                $ref: "#/components/schemas/CorrespondenceHistory"
    CreateCorrespondenceBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
          nullable: true
        type:
          type: string
          enum: [incoming, outgoing, internal]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        fromDepartmentId:
          type: integer
          nullable: true
        toDepartmentId:
          type: integer
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        createdById:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
      required:
        - subject
        - type
        - priority
    UpdateCorrespondenceBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, rejected, archived]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedToId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
    CorrespondenceHistory:
      type: object
      properties:
        id:
          type: integer
        correspondenceId:
          type: integer
        action:
          type: string
        notes:
          type: string
          nullable: true
        performedById:
          type: integer
          nullable: true
        performedByName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - correspondenceId
        - action
        - createdAt
    CreateHistoryBody:
      type: object
      properties:
        action:
          type: string
        notes:
          type: string
          nullable: true
        performedById:
          type: integer
          nullable: true
      required:
        - action
    Archive:
      type: object
      properties:
        id:
          type: integer
        correspondenceId:
          type: integer
        archiveNumber:
          type: string
        archiveLocation:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        archivedById:
          type: integer
          nullable: true
        archivedByName:
          type: string
          nullable: true
        archivedAt:
          type: string
          format: date-time
        correspondence:
          $ref: "#/components/schemas/Correspondence"
      required:
        - id
        - correspondenceId
        - archiveNumber
        - archivedAt
        - correspondence
    ArchiveCorrespondenceBody:
      type: object
      properties:
        archiveLocation:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        archivedById:
          type: integer
          nullable: true
...